moved schemas to schemas dir, merged helpers-utils
