+++ jsontableschema/schemas/geojson.json
